+++ src/Stores/ArrayStore.php
@@ -125,7 +125,7 @@ class ArrayStore implements StoreContract
     }
 
     /**
-     * Get all of the settings items.
+     * Get all the settings items.
      *
      * @return array
      */