+++ src/Stores/DatabaseStore.php
@@ -263,7 +263,7 @@ class DatabaseStore implements StoreContract
     }
 
     /**
-     * Get all of the settings items.
+     * Get all the settings items.
      *
      * @return array
      */